push api blocked ; api auth

+++ Models/GetClientInfoByUserID_Result.cs
@@ -7,7 +7,7 @@
         public string Password { get; set; }
         public string ?Profile_Url { get; set; }
         public string ?Goal { get; set; }
-        public int TargetWeight { get; set; }
+        public double TargetWeight { get; set; }
         public string ?Health_Risk { get; set; }
         public string ?Injury { get; set; }
